+++ packages/plugin-packages/mashroom-portal/type-definitions/api.ts
@@ -434,6 +434,7 @@ export type MashroomPortalApp = {
 export type MashroomAvailablePortalApp = {
     readonly name: string;
     readonly version: string;
+    readonly title: string | null | undefined;
     readonly category: string | null | undefined;
     readonly description: string | null | undefined;
     readonly tags: Array<string>;