+++ packages/plugin-packages/mashroom-portal/type-definitions/index.js
@@ -387,6 +387,7 @@ export type MashroomPortalApp = {
 export type MashroomAvailablePortalApp = {
     +name: string;
     +version: string;
+    +title: ?string;
     +category: ?string;
     +description: ?string;
     +tags: Array<string>,